requirements.txt
# Django Core
Django==5.2.6
djangorestframework==3.16.1

# Image Processing
Pillow==11.3.0

# Static Files
whitenoise==6.11.0

# CORS Headers
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.10

# Environment Variables
python-decouple==3.8
python-dotenv==1.0.0

#JWT
djangorestframework-simplejwt

# Filtering
django-filter==23.3

# Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
coverage==7.3.2

# Development Tools
ipython==8.17.2
django-extensions==3.2.3

# CSV/JSON Import
django-import-export==3.3.0

# PDF Generation
reportlab==4.0.6
xhtml2pdf==0.2.17

# Email / Notifications
django-anymail==10.2

# Authentication
django-allauth==0.54.0

# Date/Time handling
python-dateutil==2.8.2

# HTTP requests
requests==2.32.5

# JSON Web Tokens
pyjwt==2.10.1